Make the MetalServer::Restorer a dedicated class
This allows it to be reused
